fix(pacquet-cli): address dylint violations and Qodo review comments
- Replace single-letter closure params with descriptive names
- Move inline tests to external file for dylint compliance
- Fix prod/dev tri-state logic to match pnpm computeInclude
- Resolve workspace importer by relative path instead of root fallback
- Normalize importer path separators for cross-platform lockfile lookup
